Renamed namespace
Renamed the namespace for all the configurations from `Configs` to `configs`
