+++ src/infrastructure/EntityFrameworkCore/Configs/WorkItemConfiguration.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace EntityFrameworkCore.Configs;
+namespace EntityFrameworkCore.configs;
 
 public class WorkItemConfiguration : IEntityTypeConfiguration<WorkItem>
 {
