+++ src/infrastructure/EntityFrameworkCore/Configs/WorkspaceConfiguration.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace EntityFrameworkCore.Configs;
+namespace EntityFrameworkCore.configs;
 
 
 public class WorkspaceConfiguration : IEntityTypeConfiguration<Workspace>
