+++ src/infrastructure/EntityFrameworkCore/Extensions/InfrastructureExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
-namespace EntityFrameworkCore.Migrations;
+namespace EntityFrameworkCore.extensions;
 
 public static class InfrastructureExtensions
 {
